A 5-variable US business cycle data, from 1955 Q1 to 2004 Q4

A sample data to identify optimism shock.

Usage

data(optimism)

Format

A matrix and a ts object with time series of over two hundred observations on 5 variables:

productivity

quarterly factor-utilization-adjusted total factor productivity

stock_prices

quarterly end-of-period S&P 500 divided by CPI

consumption

quarterly real consumption expenditures on nondurable goods and services

real_interest_rate

quarterly real interest rate

hours_worked

quarterly hours of all persons in the non-farm business sector

The series are as described by Beaudry, Nam and Wang (2011) in section 2.2.

Source

Replication package, https://www.econometricsociety.org/publications/econometrica/2018/03/01/inference-based-structural-vector-autoregressions-identified

References

Arias, Jonas E., Juan F. Rubio‐Ramírez, and Daniel F. Waggoner. "Inference based on structural vector autoregressions identified with sign and zero restrictions: Theory and applications." Econometrica 86, no. 2 (2018): 685-720. <doi:10.3982/ECTA14468>

Beaudry, Paul, Deokwoo Nam, and Jian Wang. Do mood swings drive business cycles and is it rational?. No. w17651. National Bureau of Economic Research, 2011. <doi:10.3386/w17651>